libhsmd: Migrate handle_cannouncement_sig
